backend/requirements.txt
django==4.2.4
djangorestframework==3.14.0
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt